+++ src/main/java/nl/orlandosmits/threekidfamily/dto/ParentDto.java
@@ -1,5 +1,5 @@
 package nl.orlandosmits.threekidfamily.dto;
 
-public record ParentDto(Integer id) {
+public record ParentDto(Long id) {
 
 }
