+++ src/main/java/nl/orlandosmits/threekidfamily/dto/PartnerDto.java
@@ -1,5 +1,5 @@
 package nl.orlandosmits.threekidfamily.dto;
 
-public record PartnerDto(Integer id) {
+public record PartnerDto(Long id) {
 
 }
